Add sticky footer to page content; add full height; remove composed
