+++ src/screens/docs/index.js
@@ -4,6 +4,7 @@ import Ecology from "ecology";
 import { VictoryBar, VictoryChart, VictoryAxis, VictoryTheme, VictoryStack } from "victory";
 import Radium from "radium";
 import Prism from "prismjs";
+import { find } from "lodash";
 /* eslint-disable no-unused-vars */
 // add more language support
 import jsx from "prismjs/components/prism-jsx";
@@ -12,8 +13,9 @@ import yaml from "prismjs/components/prism-yaml";
 /* eslint-enable no-unused-vars */
 
 // Child components
+import { config } from "../../components/config";
 import InternalPage from "../../components/page-internal";
-import Markdown from "./components/markdown";
+import Markdown from "../../components/markdown";
 import TitleMeta from "../../components/title-meta";
 
 class Docs extends React.Component {
@@ -51,12 +53,13 @@ class Docs extends React.Component {
         </div>
       );
     }
+    const markdownFile = find(config, { slug: activePage }).docs;
     return (
       <Markdown
         location={this.props.location}
-        params={this.props.params}
         updateTocArray={this.updateTocArray.bind(this)}
         active={activePage}
+        markdownFile={markdownFile}
       />
     );
   }
