+++ src/styles/base/base.css
@@ -31,7 +31,7 @@ body {
   z-index: 1;
 }
 
-html, body, #content {
+html, body, #content, #content > div {
   height: 100%;
 }
