+++ src/styles/components/page.css
@@ -33,6 +33,8 @@
 }
 
 .Page-content {
+  display: flex;
+  flex-direction: column;
   flex: 1;
 }
